# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Spotlight Puerto Rico Business Directory API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses:
    get:
      operationId: listBusinesses
      tags: [businesses]
      summary: Search and list businesses
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: municipality
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of businesses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessListResponse"
    post:
      operationId: createBusiness
      tags: [businesses]
      summary: Create a new business listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessBody"
      responses:
        "201":
          description: Business created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /businesses/{id}:
    get:
      operationId: getBusiness
      tags: [businesses]
      summary: Get business by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Business detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateBusiness
      tags: [businesses]
      summary: Update a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessBody"
      responses:
        "200":
          description: Business updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteBusiness
      tags: [businesses]
      summary: Delete a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /businesses/{id}/reviews:
    get:
      operationId: listBusinessReviews
      tags: [reviews]
      summary: Get reviews for a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review for a business
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Business:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        municipality:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        logoUrl:
          type: string
        coverUrl:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        featured:
          type: boolean
        averageRating:
          type: number
        reviewCount:
          type: integer
        ownerId:
          type: string
        ownerName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, slug, status, featured, createdAt, updatedAt]
    BusinessDetail:
      allOf:
        - $ref: "#/components/schemas/Business"
        - type: object
          properties:
            hours:
              type: object
              additionalProperties:
                type: string
            socialLinks:
              type: object
              properties:
                facebook:
                  type: string
                instagram:
                  type: string
                twitter:
                  type: string
    BusinessListResponse:
      type: object
      properties:
        businesses:
          type: array
          items:
            $ref: "#/components/schemas/Business"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [businesses, total, page, limit, totalPages]
...
    CreateBusinessBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        categoryId:
          type: integer
        municipality:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        logoUrl:
          type: string
        coverUrl:
          type: string
        hours:
          type: object
          additionalProperties:
            type: string
        socialLinks:
          type: object
          properties:
            facebook:
              type: string
            instagram:
              type: string
            twitter:
              type: string
      required: [name, description, categoryId, municipality]
    UpdateBusinessBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        categoryId:
          type: integer
        municipality:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        logoUrl:
          type: string
        coverUrl:
          type: string
        hours:
          type: object
          additionalProperties:
            type: string
        socialLinks:
          type: object
          properties:
            facebook:
              type: string
            instagram:
              type: string
            twitter:
              type: string
    Review:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        userId:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
        body:
          type: string
        authorName:
          type: string
        authorImage:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, businessId, userId, rating, createdAt]
    ReviewListResponse:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
      required: [reviews, total]
    CreateReviewBody:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
        body:
          type: string
      required: [rating]
...